ml/requirements.txt
# ML Span Extractor — W31-35 Dependencies
# Install: pip install -r requirements.txt

# Core ML
torch>=2.0.0
transformers>=4.35.0
torchcrf>=1.1.0

# Classical ML baselines
scikit-learn>=1.3.0

# Evaluation
seqeval>=1.2.2

# Data handling
pyyaml>=6.0
jsonlines>=4.0.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
